+++ ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: jnivala <jnivala@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/04 09:24:41 by jnivala           #+#    #+#             */
-/*   Updated: 2020/06/04 17:14:32 by jnivala          ###   ########.fr       */
+/*   Updated: 2020/06/17 12:29:26 by jnivala          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 
 char	*ft_strdup(char *src)
 {
-	int		i;
-	char	*str;
+	size_t		i;
+	char		*str;
 
 	i = 0;
 	str = (char*)malloc(sizeof(*str) * (ft_strlen(src) + 1));
